+++ delete.cpp
@@ -1,44 +1,55 @@
 #include "include/delete.h"
 
-void deleteFile(const char *name) {
+int deleteFile(const char *name) {
 	if(isFileExists(name)) {
-		if(remove(name) == 0)
-			cout << "\nFile Deleted Successfully";
-		else
-			cout << "\nError: Cannot delete file. Please check file name!";
+		if(remove(name) == 0){
+			//File Deleted Successfully
+			return 1;
+		} else {
+			//Error: Cannot delete file. Please check file name!
+			return -1;
+		}
 	} else {
-		cout << "\nError: File not found. Please check file name!";
+		//Error: File not found. Please check file name!
+		return -2;
 	}
+	return 1;
 }
 
-void deleteDirectory(const char *dir) {
+int deleteDirectory(const char *dir) {
 
-	DIR *dp;
-	struct dirent *entry; 
-	struct stat statbuf;
+	if(isDirectoryExists(dir)) {
+		DIR *dp;
+		struct dirent *entry; 
+		struct stat statbuf;
 
-	if((dp = opendir(dir)) == NULL) { 
-		printf("Error: Unable to open directory! %s\n", dir);
-		return; 
-	}
-
-	chdir(dir);
-
-	while((entry = readdir(dp)) != NULL) {
-		lstat(entry->d_name, &statbuf); 
-		if(S_ISDIR(statbuf.st_mode)) {
-			if(strcmp(".",entry->d_name) == 0 || strcmp("..",entry->d_name) == 0)
-				continue;
-			deleteDirectory(entry->d_name);
+		if((dp = opendir(dir)) == NULL) { 
+			printf("Error: Unable to open directory! %s\n", dir);
+			return; 
 		}
-		else {
-			remove(entry->d_name);
+
+		chdir(dir);
+
+		while((entry = readdir(dp)) != NULL) {
+			lstat(entry->d_name, &statbuf); 
+			if(S_ISDIR(statbuf.st_mode)) {
+				if(strcmp(".",entry->d_name) == 0 || strcmp("..",entry->d_name) == 0)
+					continue;
+				deleteDirectory(entry->d_name);
+			}
+			else {
+				remove(entry->d_name);
+			}
 		}
+
+		chdir("..");
+
+		rmdir(dir);
+
+		closedir(dp); 
+	} else {
+		//Error: Delete Directory does not exists
+		return -1;
 	}
-
-	chdir("..");
-
-	rmdir(dir);
-
-	closedir(dp); 
+	return 1;
 }
